ecommerce_back/requirements.txt
asgiref==3.6.0
astroid==2.12.13
autopep8==1.4.4
backports.zoneinfo==0.2.1
certifi==2019.3.9
chardet==3.0.4
defusedxml==0.6.0
dill==0.3.6
Django==2.2.14
django-allauth==0.39.1
django-cors-headers==3.10.0
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-environ==0.9.0
djangorestframework==3.11.1
drf-nested-routers==0.92.5
idna==2.8
isort==5.11.3
lazy-object-proxy==1.8.0
mccabe==0.7.0
mysqlclient==2.2.6
oauthlib==3.0.1
pep8==1.7.1
Pillow==6.2.2
platformdirs==2.6.0
pycodestyle==2.5.0
pylint==2.15.9
python-decouple==3.1
python3-openid==3.1.0
pytz==2021.3
requests==2.21.0
requests-oauthlib==1.2.0
sqlparse==0.2.4
stripe==2.27.0
tomli==2.0.1
tomlkit==0.11.6
typing-extensions==4.4.0
urllib3==1.24.2
wrapt==1.14.1
djongo
celery==4.4.7
redis==4.4.0
django-celery-beat==1.4.0
flower==0.9.7
